Refactor webpack rules to also compile for electron
