+++ package.json
@@ -3,13 +3,12 @@
   "productName": "Ganache",
   "version": "1.1.0",
   "description": "Personal Blockchain for Ethereum",
-  "main": "./src/init.js",
+  "main": "./dist/main/electron.js",
   "private": true,
   "scripts": {
-    "start": "electron-forge start",
-    "start-browser": "NODE_ENV=development webpack --watch",
+    "start": "concurrently \"npm run watch-electron\" \"electron-forge start\"",
     "package": "electron-forge package",
-    "make": "electron-forge make",
+    "make": "npm run compile-electron && electron-forge make",
     "publish": "electron-forge publish",
     "lint": "echo \"No linting configured\"",
     "sign-mac": "codesign --deep --force --verbose --sign 'Developer ID Application: Timothy Coulter (58RKXWC272)' ./out/make/*.dmg",
@@ -17,7 +16,10 @@
     "build-linux": "cd ./node_modules/rabin-bindings && node-gyp rebuild && cd ../../ && npm run make",
     "build-mac": "npm run make && npm run sign-mac",
     "build-windows": "bash -c \"echo ${CERT_PASS?\\\"Need to set CERT_PASS to build on Windows\\\"} > /dev/null\" && npm run make",
-    "build-browser": "NODE_ENV=production webpack"
+    "compile-electron": "NODE_ENV=production webpack --config build/webpack.config.electron.js",
+    "compile-web": "NODE_ENV=production webpack --config build/webpack.config.web.js",
+    "watch-electron": "NODE_ENV=development webpack --watch --config build/webpack.config.electron.js",
+    "watch-web": "NODE_ENV=development webpack --watch --config build/webpack.config.web.js"
   },
   "browserslist": "chrome 50",
   "build": {
@@ -163,6 +165,7 @@
     "svg2uwptiles": "0.0.4",
     "uglifyjs-webpack-plugin": "^1.2.2",
     "webpack": "^3.11.0",
+    "webpack-merge": "^4.1.2",
     "webpack-shell-plugin": "^0.5.0"
   },
   "dependencies": {
