+++ src/Kernel/actionClient.js
@@ -7,7 +7,7 @@ if (process.env.WEBPACK_TARGET === 'web') {
   ws.addEventListener('close', () => console.log('WebSocket connection closed'))
 
   module.exports = createActionClient(ws)
-} else if (process && process.type === 'renderer') { // electron-renderer
+} else if (process.env.WEBPACK_TARGET === 'electron-renderer') { // electron-renderer
   const actionClient = require('electron').ipcRenderer
   actionClient.emit('open')
   module.exports = actionClient
