+++ src/app.html
@@ -3,16 +3,6 @@
   <head>
     <meta charset="utf-8">
     <title>Ganache</title>
-    <script>
-      document.addEventListener('dragover', event => {
-        event.preventDefault()
-        return false
-      })
-      document.addEventListener('drop', event => {
-        event.preventDefault()
-        return false
-      })
-    </script>
     <style>
       html {
         box-sizing: border-box;
@@ -30,7 +20,6 @@
     </style>
   </head>
   <body>
-    <div id="root"></div>
-    <script src="index.js"></script>
+    <div id="app"></div>
   </body>
 </html>
