--- src/app.webpack.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Ganache</title>
-    <style>
-      html {
-        box-sizing: border-box;
-
-        text-shadow: 1px 1px 1px rgba(0,0,0,.004);
-
-        font-smooth: always;
-
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-        font-family: "Fira Sans", sans-serif;
-        font-weight: 200;
-        background-color: #312c2a;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
